change compile flags, other minor cleanup
Ignore-this: 9813529ce43dd84b9d700c100a617858

darcs-hash:20100127224110-9241b-039a4ca4ec324965a70558ac73ce0b4feb49256f.gz

+++ src/haskell/Config.hs
@@ -295,7 +295,7 @@ options =
     , Option "h" ["help"]
         (NoArg (\_ -> do prg <- getProgName
                          hPutStrLn stderr (usageInfo ("Usage: " ++ prg ++ " [OPTIONS...] dta_files...")
-                            (options :: [(OptDescr (ConfigParams Double -> IO (ConfigParams Double)))]))
+                            (options :: [(OptDescr (ConfigParams Float -> IO (ConfigParams Float)))]))
                          hPutStrLn stderr digestionRuleHelp
                          exitWith ExitSuccess))
         "This help text"
